requirements.txt
python-dotenv
smolagents
litellm
requests
beautifulsoup4
notion-client